(fix) add child state support
Before, adding a child state would cause a state such as `home.list`
to suggest a url of `/homelist` and a templateUrl of `homelist.tpl.html`.
Now, `home.list` will suggest `/list` and `home-list.tpl.html`.

This will still create failing e2e tests. Users will need to update
the URL and be sure to add a ui-view element in the parent's template.