Full support of classes syntax with coffescript
I added an option to the promt, when the app is scaffolded.
If 'coffescript' is choosen as the appScript language,
the user can choose to use the classes syntax. The default is yes.

I then modified the templates for _controller.coffee, _directive.coffee,
_factory.coffee and _service.coffee, all with an condition. If the
classes option was choosen, then the template will contain a
coffeescript class. Otherwise it will generate a file with the 'old'
syntax.

Currently there is no Typescript support available. I suggest to open an
issue for that.