chore(typescript) update gulp-typescript
Before, in gulp-typescript@^2.6.0, the plugin has a heart attack and
crashes. There’s an issue with the the source code and tests creating
declaration files. A fix is creating a TypeScript source and test
project. The source project creates declaration files, while the test
project does not. This enables using later versions including 2.8.0,
which has TypeScript@1.5.0 support.