feat(options): remove support for ng-route and controller-as
Closes #184

Before, when generating new components, users could specify via
CLI to use ng-route or controller-as. Now, these options are no
longer supported for overriding the config in `.yo-rc.json`.

This is being done to encourage consistency for teams on
architecture.

If users desire to change the config from the initial scaffold,
changing the contents of `.yo-rc.json` will work for changing from
not using to controller-as to using controller-as, for example.

It is strongly encouraged to never switch routers mid-development.
ng-poly is fragile and expects to not mix routers.