feat(utils/app): retrieve app name from main module file
Closes ##192

BREAKING CHANGE

If a user renamed the Angular module in app/app-module.js, ng-poly
was not smart enough to look there and instead would only look at
`package.json` for the name.

Now, ng-poly is smart enough to look at the main module for the name.

This is a potentially breaking change if people have defined more than
one Angular module in their main module.

If the user has more than one module defined in the file, just make
sure the first module defined is the main module. Then ng-poly works
as intended.